Update config Handlebars + fichiers statique
Je n'avais pas réellement tout configuré, c'est ici corrigé :
c'est ici corrigé
* Rename de inde.js en serveur.js
* ajout du css
* ajout de l'arborescence des fichier
* config hbs
* config des fichiers statique
