+++ views/layout.hbs
@@ -1,0 +1,10 @@
+<html lang="fr">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <link rel="stylesheet" href="/css/global.css"> 
+    <body>
+        {{{ body }}}
+        {{> footer}}
+    </body>
+</html>
